Serapate dropin related methods to new files in Checkout
