+++ AdyenCheckout/Core/CheckoutCore.swift
@@ -133,47 +133,7 @@ package final class CheckoutCore: CheckoutCoreProtocol {
         return CheckoutPaymentComponent(paymentComponent: paymentComponent)
     }
 
-    package func createDropIn() -> DropInComponent? {
-        guard let paymentMethods else { return nil }
-
-        let checkoutConfiguration = configuration
-        let sessionConfiguration = session?.componentConfiguration
-        let context = adyenContext
-        let paymentComponentBuilder: DropInPaymentComponentBuilder = { paymentMethod in
-            try CheckoutComponentBuilder.build(
-                forAnyPaymentMethod: paymentMethod,
-                configuration: checkoutConfiguration,
-                sessionConfiguration: sessionConfiguration,
-                context: context
-            )
-        }
-        let dropInComponent = CheckoutComponentBuilder.buildDropIn(
-            paymentMethods: paymentMethods,
-            configuration: checkoutConfiguration,
-            context: context,
-            actionComponentConfiguration: actionComponentConfiguration,
-            managementCapability: sessionManagementCapability,
-            paymentComponentBuilder: paymentComponentBuilder
-        )
-        dropInComponent.delegate = self
-        return dropInComponent
-    }
-
     package func handle(action: Action) {
         actionHandlingComponent.handle(action)
-    }
-}
-
-private extension CheckoutCore {
-
-    var sessionManagementCapability: StoredPaymentMethodManagementCapability? {
-        guard let session, session.showRemovePaymentMethodButton else { return nil }
-
-        return StoredPaymentMethodManagementCapability { [weak session] storedPaymentMethod in
-            guard let session else {
-                throw StoredPaymentMethodRemovalError.unavailable
-            }
-            try await session.disable(storedPaymentMethod: storedPaymentMethod)
-        }
     }
 }
